Changing test structure and fixing method by method
